avoid throwing exceptions in getTalker
